Move send sequence keybindings into terminalcontrib
