+++ src/vs/workbench/contrib/terminal/browser/terminal.contribution.ts
@@ -6,20 +6,13 @@
 import { getFontSnippets } from '../../../../base/browser/fonts.js';
 import { KeyCode, KeyMod } from '../../../../base/common/keyCodes.js';
 import { Schemas } from '../../../../base/common/network.js';
-import { isIOS, isWindows } from '../../../../base/common/platform.js';
 import { URI } from '../../../../base/common/uri.js';
-import './media/terminal.css';
-import './media/terminalVoice.css';
-import './media/widgets.css';
-import './media/xterm.css';
 import * as nls from '../../../../nls.js';
-import { CONTEXT_ACCESSIBILITY_MODE_ENABLED } from '../../../../platform/accessibility/common/accessibility.js';
-import { ContextKeyExpr } from '../../../../platform/contextkey/common/contextkey.js';
 import { Extensions as DragAndDropExtensions, IDragAndDropContributionRegistry, IDraggedResourceEditorInput } from '../../../../platform/dnd/browser/dnd.js';
 import { SyncDescriptor } from '../../../../platform/instantiation/common/descriptors.js';
 import { InstantiationType, registerSingleton } from '../../../../platform/instantiation/common/extensions.js';
 import { Registry } from '../../../../platform/registry/common/platform.js';
-import { GeneralShellType, ITerminalLogService, WindowsShellType } from '../../../../platform/terminal/common/terminal.js';
+import { ITerminalLogService } from '../../../../platform/terminal/common/terminal.js';
 import { TerminalLogService } from '../../../../platform/terminal/common/terminalLogService.js';
 import { registerTerminalPlatformConfiguration } from '../../../../platform/terminal/common/terminalPlatformConfiguration.js';
 import { EditorPaneDescriptor, IEditorPaneRegistry } from '../../../browser/editor.js';
@@ -27,6 +20,14 @@ import { ViewPaneContainer } from '../../../browser/parts/views/viewPaneContaine
 import { WorkbenchPhase, registerWorkbenchContribution2 } from '../../../common/contributions.js';
 import { EditorExtensions, IEditorFactoryRegistry } from '../../../common/editor.js';
 import { IViewContainersRegistry, IViewsRegistry, Extensions as ViewContainerExtensions, ViewContainerLocation } from '../../../common/views.js';
+import { ITerminalProfileService, TERMINAL_VIEW_ID, TerminalCommandId } from '../common/terminal.js';
+import { registerColors } from '../common/terminalColorRegistry.js';
+import { registerTerminalConfiguration } from '../common/terminalConfiguration.js';
+import { terminalStrings } from '../common/terminalStrings.js';
+import './media/terminal.css';
+import './media/terminalVoice.css';
+import './media/widgets.css';
+import './media/xterm.css';
 import { RemoteTerminalBackendContribution } from './remoteTerminalBackend.js';
 import { ITerminalConfigurationService, ITerminalEditorService, ITerminalGroupService, ITerminalInstanceService, ITerminalService, TerminalDataTransfers, terminalEditorId } from './terminal.js';
 import { registerTerminalActions } from './terminalActions.js';
@@ -43,14 +44,8 @@ import { TerminalMainContribution } from './terminalMainContribution.js';
 import { setupTerminalMenus } from './terminalMenus.js';
 import { TerminalProfileService } from './terminalProfileService.js';
 import { TerminalService } from './terminalService.js';
-import { TerminalViewPane } from './terminalView.js';
-import { ITerminalProfileService, TERMINAL_VIEW_ID, TerminalCommandId } from '../common/terminal.js';
-import { registerColors } from '../common/terminalColorRegistry.js';
-import { registerTerminalConfiguration } from '../common/terminalConfiguration.js';
-import { TerminalContextKeyStrings, TerminalContextKeys } from '../common/terminalContextKey.js';
-import { terminalStrings } from '../common/terminalStrings.js';
-import { registerSendSequenceKeybinding } from './terminalKeybindings.js';
 import { TerminalTelemetryContribution } from './terminalTelemetry.js';
+import { TerminalViewPane } from './terminalView.js';
 
 // Register services
 registerSingleton(ITerminalLogService, TerminalLogService, InstantiationType.Delayed);
@@ -133,119 +128,7 @@ Registry.as<IViewsRegistry>(ViewContainerExtensions.ViewsRegistry).registerViews
 	}
 }], VIEW_CONTAINER);
 
-// Register actions
 registerTerminalActions();
-
-const enum Constants {
-	/** The text representation of `^<letter>` is `'A'.charCodeAt(0) + 1`. */
-	CtrlLetterOffset = 64
-}
-
-// An extra Windows-only ctrl+v keybinding is used for pwsh that sends ctrl+v directly to the
-// shell, this gets handled by PSReadLine which properly handles multi-line pastes. This is
-// disabled in accessibility mode as PowerShell does not run PSReadLine when it detects a screen
-// reader. This works even when clipboard.readText is not supported.
-if (isWindows) {
-	registerSendSequenceKeybinding(String.fromCharCode('V'.charCodeAt(0) - Constants.CtrlLetterOffset), { // ctrl+v
-		when: ContextKeyExpr.and(TerminalContextKeys.focus, ContextKeyExpr.equals(TerminalContextKeyStrings.ShellType, GeneralShellType.PowerShell), CONTEXT_ACCESSIBILITY_MODE_ENABLED.negate()),
-		primary: KeyMod.CtrlCmd | KeyCode.KeyV
-	});
-}
-
-// Map certain keybindings in pwsh to unused keys which get handled by PSReadLine handlers in the
-// shell integration script. This allows keystrokes that cannot be sent via VT sequences to work.
-// See https://github.com/microsoft/terminal/issues/879#issuecomment-497775007
-registerSendSequenceKeybinding('\x1b[24~a', { // F12,a -> ctrl+space (MenuComplete)
-	when: ContextKeyExpr.and(TerminalContextKeys.focus, ContextKeyExpr.equals(TerminalContextKeyStrings.ShellType, GeneralShellType.PowerShell), TerminalContextKeys.terminalShellIntegrationEnabled, CONTEXT_ACCESSIBILITY_MODE_ENABLED.negate()),
-	primary: KeyMod.CtrlCmd | KeyCode.Space,
-	mac: { primary: KeyMod.WinCtrl | KeyCode.Space }
-});
-registerSendSequenceKeybinding('\x1b[24~b', { // F12,b -> alt+space (SetMark)
-	when: ContextKeyExpr.and(TerminalContextKeys.focus, ContextKeyExpr.equals(TerminalContextKeyStrings.ShellType, GeneralShellType.PowerShell), TerminalContextKeys.terminalShellIntegrationEnabled, CONTEXT_ACCESSIBILITY_MODE_ENABLED.negate()),
-	primary: KeyMod.Alt | KeyCode.Space
-});
-registerSendSequenceKeybinding('\x1b[24~c', { // F12,c -> shift+enter (AddLine)
-	when: ContextKeyExpr.and(TerminalContextKeys.focus, ContextKeyExpr.equals(TerminalContextKeyStrings.ShellType, GeneralShellType.PowerShell), TerminalContextKeys.terminalShellIntegrationEnabled, CONTEXT_ACCESSIBILITY_MODE_ENABLED.negate()),
-	primary: KeyMod.Shift | KeyCode.Enter
-});
-registerSendSequenceKeybinding('\x1b[24~d', { // F12,d -> shift+end (SelectLine) - HACK: \x1b[1;2F is supposed to work but it doesn't
-	when: ContextKeyExpr.and(TerminalContextKeys.focus, ContextKeyExpr.equals(TerminalContextKeyStrings.ShellType, GeneralShellType.PowerShell), TerminalContextKeys.terminalShellIntegrationEnabled, CONTEXT_ACCESSIBILITY_MODE_ENABLED.negate()),
-	mac: { primary: KeyMod.Shift | KeyMod.CtrlCmd | KeyCode.RightArrow }
-});
-
-// Always on pwsh keybindings
-registerSendSequenceKeybinding('\x1b[1;2H', { // Shift+home
-	when: ContextKeyExpr.and(TerminalContextKeys.focus, ContextKeyExpr.equals(TerminalContextKeyStrings.ShellType, GeneralShellType.PowerShell)),
-	mac: { primary: KeyMod.Shift | KeyMod.CtrlCmd | KeyCode.LeftArrow }
-});
-
-// Map ctrl+alt+r -> ctrl+r when in accessibility mode due to default run recent command keybinding
-registerSendSequenceKeybinding('\x12', {
-	when: ContextKeyExpr.and(TerminalContextKeys.focus, CONTEXT_ACCESSIBILITY_MODE_ENABLED),
-	primary: KeyMod.CtrlCmd | KeyMod.Alt | KeyCode.KeyR,
-	mac: { primary: KeyMod.WinCtrl | KeyMod.Alt | KeyCode.KeyR }
-});
-
-// Map ctrl+alt+g -> ctrl+g due to default go to recent directory keybinding
-registerSendSequenceKeybinding('\x07', {
-	when: TerminalContextKeys.focus,
-	primary: KeyMod.CtrlCmd | KeyMod.Alt | KeyCode.KeyG,
-	mac: { primary: KeyMod.WinCtrl | KeyMod.Alt | KeyCode.KeyG }
-});
-
-// send ctrl+c to the iPad when the terminal is focused and ctrl+c is pressed to kill the process (work around for #114009)
-if (isIOS) {
-	registerSendSequenceKeybinding(String.fromCharCode('C'.charCodeAt(0) - Constants.CtrlLetterOffset), { // ctrl+c
-		when: ContextKeyExpr.and(TerminalContextKeys.focus),
-		primary: KeyMod.WinCtrl | KeyCode.KeyC
-	});
-}
-
-// Delete word left: ctrl+w
-registerSendSequenceKeybinding(String.fromCharCode('W'.charCodeAt(0) - Constants.CtrlLetterOffset), {
-	primary: KeyMod.CtrlCmd | KeyCode.Backspace,
-	mac: { primary: KeyMod.Alt | KeyCode.Backspace }
-});
-if (isWindows) {
-	// Delete word left: ctrl+h
-	// Windows cmd.exe requires ^H to delete full word left
-	registerSendSequenceKeybinding(String.fromCharCode('H'.charCodeAt(0) - Constants.CtrlLetterOffset), {
-		when: ContextKeyExpr.and(TerminalContextKeys.focus, ContextKeyExpr.equals(TerminalContextKeyStrings.ShellType, WindowsShellType.CommandPrompt)),
-		primary: KeyMod.CtrlCmd | KeyCode.Backspace,
-	});
-}
-// Delete word right: alt+d [27, 100]
-registerSendSequenceKeybinding('\u001bd', {
-	primary: KeyMod.CtrlCmd | KeyCode.Delete,
-	mac: { primary: KeyMod.Alt | KeyCode.Delete }
-});
-// Delete to line start: ctrl+u
-registerSendSequenceKeybinding('\u0015', {
-	mac: { primary: KeyMod.CtrlCmd | KeyCode.Backspace }
-});
-// Move to line start: ctrl+A
-registerSendSequenceKeybinding(String.fromCharCode('A'.charCodeAt(0) - 64), {
-	mac: { primary: KeyMod.CtrlCmd | KeyCode.LeftArrow }
-});
-// Move to line end: ctrl+E
-registerSendSequenceKeybinding(String.fromCharCode('E'.charCodeAt(0) - 64), {
-	mac: { primary: KeyMod.CtrlCmd | KeyCode.RightArrow }
-});
-// NUL: ctrl+shift+2
-registerSendSequenceKeybinding('\u0000', {
-	primary: KeyMod.CtrlCmd | KeyMod.Shift | KeyCode.Digit2,
-	mac: { primary: KeyMod.WinCtrl | KeyMod.Shift | KeyCode.Digit2 }
-});
-// RS: ctrl+shift+6
-registerSendSequenceKeybinding('\u001e', {
-	primary: KeyMod.CtrlCmd | KeyMod.Shift | KeyCode.Digit6,
-	mac: { primary: KeyMod.WinCtrl | KeyMod.Shift | KeyCode.Digit6 }
-});
-// US (Undo): ctrl+/
-registerSendSequenceKeybinding('\u001f', {
-	primary: KeyMod.CtrlCmd | KeyCode.Slash,
-	mac: { primary: KeyMod.WinCtrl | KeyCode.Slash }
-});
 
 setupTerminalCommands();
